requirements.txt
fastapi
uvicorn
python-multipart
fastapi-mail
python-dotenv
pydantic
jinja2